pyproject.toml
[tool.poetry]
name = "market"
version = "0.1.0"
description = ""
authors = []

[tool.poetry.dependencies]
python = "3.11.4"
django = "^4.1.6"
django-cachalot = "^2.5.3"
djangorestframework = "^3.14.0"
aiogram = "^2.25.1"
gunicorn = "^20.1.0"
python-dotenv = "^1.0.0"
psycopg2-binary = "^2.9.5"
pillow = "^9.4.0"
rollbar = "^0.16.3"
redis = "^4.5.5"
django-redis = "^5.3.0"

[tool.poetry.dev-dependencies]
flake8 = "^6.0.0"
flake8-bugbear = "^23.3.12"
flake8-builtins = "^2.1.0"
flake8-cognitive-complexity = "^0.1.0"
flake8-commas = "^2.1.0"
flake8-django = "^1.1.5"
flake8-quotes = "^3.3.2"
flake8-todo = "^0.7"
flake8-variables-names = "^0.0.5"
autoflake = "^1.3.1"
django-silk = "^5.0.3"
django_debug_toolbar  = "^3.8.1"
isort = "^5.12.0"
